pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glyphcast"
version = "0.1.0"
description = "GPU-accelerated ASCII conversion for videos and GIFs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "aringadre76" }
]
dependencies = [
  "numpy>=1.23",
  "opencv-python>=4.10",
  "pillow>=10.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
  "typer>=0.12",
  "tqdm>=4.66",
]

[project.optional-dependencies]
train = [
  "joblib>=1.4",
  "scikit-learn>=1.5",
  "torch>=2.3",
  "torchvision>=0.18",
]
edge = [
  "kornia>=0.8",
  "torch>=2.3",
  "torchvision>=0.18",
]
dev = [
  "mypy>=1.10",
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
glyphcast = "glyphcast.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
include = ["glyphcast*"]
